requirements.txt
absl-py==1.3.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.14.1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
babel==2.16.0
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
black==23.12.1
bleach==6.1.0
blinker==1.9.0
boto3==1.36.26
botocore==1.36.26
braceexpand==0.1.7
Cartopy==0.24.1
cattrs==23.2.3
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
click==8.1.7
cmake==3.31.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
coverage==7.6.12
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.11
datasets==3.3.1
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
diffdist==0.1
dill==0.3.8
distlib==0.3.9
docker-pycreds==0.4.0
docopt-ng==0.9.0
docutils==0.16
exceptiongroup==1.2.1
executing==2.1.0
faiss-gpu==1.7.2
fastapi==0.115.8
fastjsonschema==2.20.0
ffmpy==0.5.0
filelock==3.16.1
filetype==1.2.0
flake8==6.1.0
Flask==3.1.0
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.10.0
ftfy==6.2.0
gdown==5.2.0
geojson-rewind==1.1.0
geomet==1.1.0
geos==0.2.3
git-filter-repo==2.47.0
gitdb==4.0.11
GitPython==3.1.43
gradio==4.44.1
gradio_client==1.3.0
greenlet==3.1.1
grpcio==1.64.1
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.29.0
identify==2.6.1
idna==3.10
imageio==2.36.0
imagesize==1.4.1
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jedi==0.19.1
jeepney==0.8.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kaggle==1.6.14
keyring==25.5.0
Kivy==2.3.1
Kivy-Garden==0.1.5
kiwisolver==1.4.7
lazy_loader==0.4
libmr==0.1.9
lit==18.1.8
loguru==0.7.2
lxml==5.3.0
Mako==1.3.9
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.3
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-boto3-s3==1.36.21
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nodeenv==1.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
open_clip_torch==2.26.1
opencv-python==4.10.0.84
orjson==3.10.15
overrides==7.7.0
packaging==24.1
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pathtools==0.1.2
pexpect==4.9.0
Pillow==9.5.0
platformdirs==4.3.6
pluggy==1.5.0
plyer==2.1.0
pre_commit==4.0.1
prettytable==3.10.0
prometheus_client==0.21.0
promise==2.3
prompt_toolkit==3.0.48
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pydub==0.25.1
pyflakes==3.1.0
pygbif==0.6.4
Pygments==2.18.0
pyinaturalist==0.19.0
pyobjus==1.2.3
pyparsing==3.2.0
pyproj==3.7.0
pyrate-limiter==2.10.0
pyshp==2.3.1
PySocks==1.7.1
pytest==8.3.4
pytest-asyncio==0.21.2
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-dwca-reader==0.15.1
python-forge==18.6.0
python-json-logger==2.0.7
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0
pyzmq==26.2.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-cache==1.2.1
requests-ratelimiter==0.7.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
rsa==4.7.2
ruff==0.9.7
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
SecretStorage==3.3.3
segment-anything==1.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentry-sdk==1.45.1
setproctitle==1.3.3
shapely==2.0.6
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
SQLAlchemy-Utils==0.40.0
stack-data==0.6.3
starlette==0.45.3
structlog==22.3.0
sympy==1.13.1
tensorboard==2.17.0
tensorboard-data-server==0.7.2
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.4.0
tokenizers==0.19.1
tomli==2.0.2
tomlkit==0.12.0
torch==2.1.2
torchmetrics==0.11.4
torchsummary==1.5.1
torchvision==0.16.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
triton==2.1.0
typer==0.12.5
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.27.1
wandb==0.17.9
wcwidth==0.2.13
webcolors==24.8.0
webdataset==0.2.100
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.13
xxhash==3.4.1
yarl==1.9.4
zipp==3.20.2
